ARM: Tegra: Dalmore: Fix Shifting in dalmore_wakeup_key
The correct wakes were not being detected due to an issue
in shifting by values greater than 32.

dalmore_wakeup_key will now correctly determine which GPIO
caused the WAKE.

Bug 1248964

Reviewed-on: http://git-master/r/209705
(cherry picked from commit e48e943cfcc331a53a33806cad112f18a08f57f7)
Reviewed-on: http://git-master/r/226431

+++ arch/arm/mach-tegra/board-dalmore-kbc.c
@@ -45,9 +45,9 @@ static int dalmore_wakeup_key(void)
 		| (u64)readl(IO_ADDRESS(TEGRA_PMC_BASE)
 		+ PMC_WAKE2_STATUS) << 32;
 
-	if (status & ((u64)1 << TEGRA_WAKE_GPIO_PQ0))
+	if (status & (1ULL << TEGRA_WAKE_GPIO_PQ0))
 		wakeup_key = KEY_POWER;
-	else if (status & ((u64)1 << TEGRA_WAKE_GPIO_PS0))
+	else if (status & (1ULL << TEGRA_WAKE_GPIO_PS0))
 		wakeup_key = SW_LID;
 	else
 		wakeup_key = KEY_RESERVED;
